requirements.txt
discord.py
python-dotenv
openai
google-generativeai
requests
aiohttp